+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 * Auto generation of row edit form.
 * Initial implementation of virtual scrolling.
 * Infinite scrolling.
+* Perfomance improvements.
 * Validation framework.
 * Persistence plugin.
 * Email, url cell editor.
@@ -16,7 +17,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ## [5.2.4] - 2018-02-16
 ### Changed
 - Renamed `action-bar-core` to `action-bar`.
-- 
+- `layout="row"` refactored to `class="layout-row"`
+
+### Fixed
+- Url editor.
+- Text alingment in the file editor.
+- Navigation should trigger detect changes.
 
 ## [5.2.3] - 2018-02-12
 ### Fixed